// src/app/modules/fianzas/pages/core/helpers/enums/fnz.notificacion.enum.ts
export enum EFnzNotificacion {
    FORMATOS_FIRMADOS = 1,
    SOLICITUD_FIRMA_ASEGURADO = 2,
    SOLICITUD_FIRMA_EJECUTIVO = 3,
    SOLICITUD = 4,
    CONFIRMACION_ENTREGA = 5,
    REVISION_DOCUMENTAL = 6,
    DATOS_CONTACTO = 7,
    NO_CONTINUA_PROCESO = 8,
    RECEPCION_DOCUMENTOS = 9,
    RECOLECCION_FISICOS = 10,
    FISICOS_ENVIADOS = 11,
    REVISION_FORMATOS = 12,
    FIRMA_DOCUMENTOS = 13,
    REVISION_DOCUMENTACION_FISICA = 14,
    GENERACION_DE_FORMATOS = 15
}